requirements.txt
matplotlib==3.4.2
networkx==2.5.1
numpy
pillow
pyparsing==2.4.7
six==1.16.0
open3d
opencv-python==4.5.2.52
pandas
pygame
scikit-learn==0.24.2
scipy==1.6.3
seaborn
lxml
shapely
omegaconf
hydra-core
pynput